benchmark: create some tree functions
